rename layout components and change folder structure
